fix compiler waring (regarding size_t and 64bit architecture)
git-svn-id: svn://svn.minisip.org/minisip/trunk@2193 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libminisip/source/sip/DefaultDialogHandler.cxx
@@ -953,7 +953,7 @@ void DefaultDialogHandler::sendIMOk(MRef<SipRequest*> bye, const string &branch)
 
 void DefaultDialogHandler::sendIM(const string &branch, string msg, int im_seq_no, string toUri){
 	
-	uint32_t posAt;
+	size_t posAt;
 	
 	posAt = toUri.find("@");
 	if( posAt == string::npos ) { //toUri does not have a domain ...
